+++ R/data-pop_est.R
@@ -1,4 +1,4 @@
-#' Dataframe of annual population counts by age and race, 1979-2015
+#' Dataframe of annual population counts by age and race, 1979-2020
 #'
 #' A data set containing annual population counts by age (5-year bins,
 #' 18 groups), sex, and race. All data are from the US Census Bureau Population
@@ -6,9 +6,9 @@
 #'
 #' @docType data
 #'
-#' @format A data frame with 7992 rows and 6 columns
+#' @format A data frame with 9072 rows and 6 columns
 #' \describe{
-#'   \item{year}{year of observation (int)}
+#'   \item{year}{year of observation}
 #'   \item{age}{starting age for that age category}
 #'   \item{age_cat}{age category in human readable form}
 #'   \item{sex}{male, female, or both}