+++ R/download_all_pop_data.R
@@ -1,6 +1,6 @@
-#' Download 1979-2015 Population Estimates
+#' Download 1979-2020 Population Estimates
 #'
-#' All annual population estimates from 1979 to 2015 by age (18 bins) and
+#' All annual population estimates from 1979 to 2020 by age (18 bins) and
 #' race. This is the function used to create the pop_est dataframe.
 #'
 #' @return Dataframe with US Census Bureau population estimates